refactor: 함수들 관련 파일로 이동
