+++ src/components/columns.js
@@ -1,4 +1,5 @@
 import { createColumnTemplate } from "./templates.js";
+import Card from "../components/cards.js";
 
 // Column element
 export default function Column({ title, id, count }) {
@@ -9,4 +10,17 @@ export default function Column({ title, id, count }) {
   column.innerHTML = createColumnTemplate(title, id, count);
 
   return column;
+}
+
+// Column의 '+' 버튼 클릭시 카드 추가 함수
+export function addCard({ target }) {
+  const column = target.closest(".column");
+  const cardList = column.querySelector("#cardList");
+  const isExistCard = column.querySelector(".newCard");
+  if (!isExistCard) {
+    const card = Card();
+    cardList.insertAdjacentElement("afterbegin", card);
+  } else {
+    isExistCard.remove();
+  }
 }
